requirements.txt
langchain==0.3.4
langchain-chroma==0.1.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
openai==1.52.0
chromadb==0.5.15
sentence-transformers
streamlit
python-dotenv
requests==2.32.3
pdfminer.six==20240706